devop: imp bubble sort
